Move SharedFilesDialog to /gui/FileTransfer/
And remove old DEFINE because very olds files are deleted.
