+++ retroshare-gui/src/gui/FileTransfer/SharedFilesDialog.ui
