+++ retroshare-gui/src/retroshare-gui.pro
@@ -351,7 +351,6 @@ HEADERS +=  rshare.h \
             gui/RemoteDirModel.h \
             gui/RetroShareLink.h \
             gui/SearchTreeWidget.h \
-            gui/SharedFilesDialog.h \
             gui/ShareManager.h \
 #            gui/ShareDialog.h \
 #            gui/SFListDelegate.h \
@@ -360,6 +359,7 @@ HEADERS +=  rshare.h \
             gui/LogoBar.h \
             gui/common/AvatarDialog.h \
             gui/FileTransfer/SearchDialog.h \
+            gui/FileTransfer/SharedFilesDialog.h \
             gui/FileTransfer/xprogressbar.h \
             gui/FileTransfer/DetailsDialog.h \
             gui/FileTransfer/FileTransferInfoWidget.h \
@@ -595,11 +595,11 @@ FORMS +=    gui/StartDialog.ui \
             gui/FileTransfer/TransfersDialog.ui \
             gui/FileTransfer/DetailsDialog.ui \
             gui/FileTransfer/SearchDialog.ui \
+            gui/FileTransfer/SharedFilesDialog.ui \
             gui/MainWindow.ui \
             gui/NetworkView.ui \
             gui/MessengerWindow.ui \
             gui/FriendsDialog.ui \
-            gui/SharedFilesDialog.ui \
             gui/ShareManager.ui \
 #            gui/ShareDialog.ui \
             gui/MessagesDialog.ui \
@@ -723,7 +723,6 @@ SOURCES +=  main.cpp \
             gui/RsAutoUpdatePage.cpp \
             gui/RetroShareLink.cpp \
             gui/SearchTreeWidget.cpp \
-            gui/SharedFilesDialog.cpp \
             gui/ShareManager.cpp \
 #            gui/ShareDialog.cpp \
 #            gui/SFListDelegate.cpp \
@@ -735,6 +734,7 @@ SOURCES +=  main.cpp \
             gui/help/browser/helpbrowser.cpp \
             gui/help/browser/helptextbrowser.cpp \
             gui/FileTransfer/SearchDialog.cpp \
+            gui/FileTransfer/SharedFilesDialog.cpp \
             gui/FileTransfer/TransfersDialog.cpp \
             gui/FileTransfer/FileTransferInfoWidget.cpp \
             gui/FileTransfer/DLListDelegate.cpp \
@@ -995,51 +995,6 @@ pluginmgr {
         DEFINES *= PLUGINMGR
 
 }
-
-#blogs {
-#
-#DEPENDPATH += gui/unfinished \
-#
-#HEADERS += gui/unfinished/blogs/BlogsDialog.h \
-#           gui/unfinished/blogs/CreateBlog.h \
-#           gui/unfinished/blogs/CreateBlogMsg.h \
-#           gui/unfinished/blogs/BlogsMsgItem.h \
-#           gui/unfinished/blogs/BlogDetails.h \
-#           gui/feeds/BlogNewItem.h \
-#           gui/feeds/BlogMsgItem.h \
-#
-#FORMS += gui/unfinished/blogs/BlogsDialog.ui \
-#         gui/unfinished/blogs/CreateBlog.ui \
-#         gui/unfinished/blogs/CreateBlogMsg.ui \
-#         gui/unfinished/blogs/BlogsMsgItem.ui \
-#         gui/unfinished/blogs/BlogDetails.ui \
-#         gui/feeds/BlogNewItem.ui \
-#         gui/feeds/BlogMsgItem.ui \
-#
-#SOURCES += gui/unfinished/blogs/BlogsDialog.cpp \
-#           gui/unfinished/blogs/CreateBlog.cpp \
-#           gui/unfinished/blogs/CreateBlogMsg.cpp \
-#           gui/unfinished/blogs/BlogsMsgItem.cpp \
-#           gui/unfinished/blogs/BlogDetails.cpp \
-#           gui/feeds/BlogNewItem.cpp \
-#           gui/feeds/BlogMsgItem.cpp \
-#
-#DEFINES += BLOGS
-#}
-
-# use_links {
-# HEADERS += gui/AddLinksDialog.h \
-#            gui/LinksDialog.h
-# 
-# FORMS += gui/AddLinksDialog.ui \
-#          gui/LinksDialog.ui
-# 
-# SOURCES += gui/AddLinksDialog.cpp \
-#            gui/LinksDialog.cpp
-# 
-# DEFINES += RS_USE_LINKS
-# }
-
 
 idle {
